Refactor MockStreamSourceContext
